ndsecure_backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==3.2.25
django-cors-headers==4.1.0
django-environ==0.11.2
djangorestframework==3.15.1
firebase-admin==6.5.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.62.3
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib-metadata==6.7.0
msgpack==1.0.5
packaging==24.0
proto-plus==1.24.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.4
python-dotenv==0.21.1
python-json-logger==2.0.7
pytz==2024.2
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.7
zipp==3.15.0
